tests(stern): Move tests client to dedicated library
