+++ lib/stern/tests/fixture/test_fixture.h
